requirements.txt
numpy
opencv-python-headless
streamlit
deepface
av
Pillow
streamlit-webrtc
